requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
weasyprint==61.2
markdown==3.6
bleach==6.1.0
requests==2.31.0
django-filter==24.2
